Обновление .Net Core 3.1.2, перевод комментариев

+++ Novartment.Base.Net.Mime/Net/Mime/LexicalElements/EncodingService/HeaderDecoder.cs
@@ -134,7 +134,7 @@ namespace Novartment.Base.Net.Mime
 						while (pos < source.Length)
 						{
 							octet = source[pos];
-							if ((octet >= asciiClasses.Length) || ((asciiClasses[octet] & (short)AsciiCharClasses.Visible) == 0))
+							if ((octet >= asciiClasses.Length) || ((asciiClasses[octet] & AsciiCharClasses.Visible) == 0))
 							{
 								break;
 							}
