+++ Novartment.Base/Text/EstimatingEncoders/QuotedStringEstimatingEncoder.cs
@@ -67,7 +67,7 @@ namespace Novartment.Base.Text
 			while ((srcPos < source.Length) && (dstPos < maxOutCount))
 			{
 				var octet = source[srcPos];
-				var isEnabledClass = (octet < asciiClasses.Length) && ((asciiClasses[octet] & (short)_enabledClasses) != 0);
+				var isEnabledClass = (octet < asciiClasses.Length) && ((asciiClasses[octet] & _enabledClasses) != 0);
 				if (!isEnabledClass)
 				{
 					break;
@@ -125,7 +125,7 @@ namespace Novartment.Base.Text
 			while ((srcPos < source.Length) && (outOffset < maxOutCount))
 			{
 				var octet = source[srcPos];
-				var isEnabledClass = (octet < asciiClasses.Length) && ((asciiClasses[octet] & (short)_enabledClasses) != 0);
+				var isEnabledClass = (octet < asciiClasses.Length) && ((asciiClasses[octet] & _enabledClasses) != 0);
 				if (!isEnabledClass)
 				{
 					break;
